Deprecate Youtube Downloader Module
The YouTube API has undergone policy changes (July 2024) that restrict the functionality of the YouTube Downloader module in localkit.  This module is now deprecated.

The changes in the API make it difficult for programs like localkit to interact with YouTube effectively.

This commit removes the YouTube Downloader module and provides a clear deprecation warning to developers.

Signed-off-by: Jesus Graterol <jesusgraterol.dev@protonmail.com>